improvement(Addressed PR feedback):
