+++ src/test/java/org/opentripplanner/middleware/testutils/OtpTestUtils.java
@@ -71,8 +71,8 @@ public class OtpTestUtils {
     public static final OtpDispatcherResponse REROUTE_PLAN_RESPONSE =
         initializeMockPlanResponse("otp/response/rerouteResponse.json");
 
-    public static final OtpDispatcherResponse OTP2_DISPATCHER_PLAN_RESPONSE_TRIP_NOT_POSSIBLE_AT_MIDNIGHT =
-        initializeMockPlanResponse("otp/response/trip-not-possible-at-midnight.json");
+    public static final OtpDispatcherResponse OTP2_DISPATCHER_PLAN_RESPONSE_TRIP_QUERIED_AT_MIDNIGHT =
+        initializeMockPlanResponse("otp/response/trip-queried-at-midnight.json");
 
     /**
      * Prevents the mock OTP server from being initialized more than once
