+++ src/test/resources/org/opentripplanner/middleware/otp/response/trip-queried-at-midnight.json
